+++ workers/ohsome_quality_analyst/geodatabase/create_results_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS results (
     fid text,
     timestamp_oqt timestamp with time zone DEFAULT CURRENT_TIMESTAMP,
     timestamp_osm timestamp with time zone,
-    result_label text,
+    result_class integer,
     result_value float,  -- VALUE is an SQL keyword
     result_description text,
     result_svg text,