+++ workers/ohsome_quality_analyst/indicators/ghs_pop_comparison_roads/indicator.py
@@ -25,7 +25,6 @@ class GhsPopComparisonRoads(BaseIndicator):
         self.area = None
         self.pop_count_per_sqkm = None
         self.feature_length = None
-        self.feature_length_per_sqkm = None
 
     @classmethod
     def attribution(cls) -> str:
@@ -59,15 +58,15 @@ class GhsPopComparisonRoads(BaseIndicator):
         self.feature_length = query_results["result"][0]["value"] / 1000
         timestamp = query_results["result"][0]["timestamp"]
         self.result.timestamp_osm = dateutil.parser.isoparse(timestamp)
-        self.feature_length_per_sqkm = self.feature_length / self.area
-        self.pop_count_per_sqkm = self.pop_count / self.area
 
     def calculate(self) -> None:
+        self.pop_count_per_sqkm = self.pop_count / self.area
+        self.result.value = self.feature_length / self.area  # feature_length_per_sqkm
         description = Template(self.metadata.result_description).substitute(
             pop_count=round(self.pop_count),
             area=round(self.area, 1),
             pop_count_per_sqkm=round(self.pop_count_per_sqkm, 1),
-            feature_length_per_sqkm=round(self.feature_length_per_sqkm, 1),
+            feature_length_per_sqkm=round(self.result.value, 1),
         )
 
         green_road_density = self.green_threshold_function(self.pop_count_per_sqkm)
@@ -76,27 +75,24 @@ class GhsPopComparisonRoads(BaseIndicator):
         if self.pop_count_per_sqkm == 0:
             return
         # road density is compliant to the green values or even higher
-        elif self.feature_length_per_sqkm >= green_road_density:
-            self.result.value = 1.0
+        elif self.result.value >= green_road_density:
+            self.result.class_ = 5
             self.result.description = (
                 description + self.metadata.label_description["green"]
             )
-            self.result.label = "green"
-        # road density is too small, none, or too short roads
-        elif self.feature_length_per_sqkm < yellow_road_density:
-            self.result.value = 0.0
-            self.result.description = (
-                description + self.metadata.label_description["red"]
-            )
-            self.result.label = "red"
         # road density is compliant to the yellow values
         # we assume there could be more roads mapped
-        else:
-            self.result.value = 0.5
+        elif self.result.value >= yellow_road_density:
+            self.result.class_ = 3
             self.result.description = (
                 description + self.metadata.label_description["yellow"]
             )
-            self.result.label = "yellow"
+        # road density is too small, none, or too short roads
+        else:
+            self.result.class_ = 1
+            self.result.description = (
+                description + self.metadata.label_description["red"]
+            )
 
     def create_figure(self) -> None:
         if self.result.label == "undefined":
@@ -143,7 +139,7 @@ class GhsPopComparisonRoads(BaseIndicator):
         ax.fill_between(
             x,
             y1,
-            max(max(y1), self.feature_length_per_sqkm),
+            max(max(y1), self.result.value),
             alpha=0.5,
             color="green",
         )
@@ -151,7 +147,7 @@ class GhsPopComparisonRoads(BaseIndicator):
         # Plot pont as circle ("o").
         ax.plot(
             self.pop_count_per_sqkm,
-            self.feature_length_per_sqkm,
+            self.result.value,
             "o",
             color="black",
             label="location",